Question_improver/conda_requirements.txt
pandas==2.2.1
scikit-learn==1.3.0
numpy==1.23.5
seaborn==0.12.2
matplotlib==3.8.0
lime==0.2.0.1
umap-learn==0.5.4
spacy==3.7.2
tqdm==4.65.0
nltk==3.8.1